feat: rename server.ts to app.ts
